haiti: compatibilidad con facturacion_base 108

+++ controller/ventas_megafacturador.php
@@ -11,11 +11,12 @@
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
+ * GNU Affero General Public License for more details.
  *
- * You should have received a copy of the GNU Lesser General Public License
+ * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+require_model('articulo.php');
 require_model('albaran_cliente.php');
 require_model('pedido_cliente.php');
 require_model('albaran_proveedor.php');
@@ -64,6 +65,7 @@ class ventas_megafacturador extends fs_controller {
     public $lista_albaranes_pendientes;
     public $lista_albaranes_pendientes_total;
     private $asiento_factura;
+    private $articulos;
     private $cliente;
     private $ejercicio;
     private $ejercicios;
@@ -78,6 +80,7 @@ class ventas_megafacturador extends fs_controller {
     }
 
     protected function private_core() {
+        $this->articulos = new articulo();
         $this->asiento_factura = new asiento_factura();
         $this->cliente = new cliente();
         $this->ejercicio = new ejercicio();
@@ -149,8 +152,10 @@ class ventas_megafacturador extends fs_controller {
      * generados en base a las series de los pedidos y las fechas elegidas
      */
     private function generar_albaranes() {
-        $total = $this->total_pedidos_pendientes()['total'];
+        $total_pedidos = $this->total_pedidos_pendientes();
+        $total = $total_pedidos['total'];
         $contador = 0;
+        $errores = 0;
         foreach ($this->pedidos_pendientes() as $pedido) {
             $continuar = FALSE;
             $albaran = new albaran_cliente();
@@ -170,17 +175,17 @@ class ventas_megafacturador extends fs_controller {
             $albaran->codpostal = $pedido->codpostal;
             $albaran->codserie = $pedido->codserie;
             $albaran->direccion = $pedido->direccion;
-            $albaran->neto = $pedido->neto;
             $albaran->nombrecliente = $pedido->nombrecliente;
             $albaran->observaciones = $pedido->observaciones;
             $albaran->provincia = $pedido->provincia;
-            $albaran->total = $pedido->total;
-            $albaran->totaliva = $pedido->totaliva;
             $albaran->numero2 = $pedido->numero2;
-            $albaran->irpf = $pedido->irpf;
             $albaran->porcomision = $pedido->porcomision;
-            $albaran->totalirpf = $pedido->totalirpf;
-            $albaran->totalrecargo = $pedido->totalrecargo;
+            $albaran->neto = 0;
+            $albaran->total = 0;
+            $albaran->totaliva = 0;
+            $albaran->irpf = 0;
+            $albaran->totalirpf = 0;
+            $albaran->totalrecargo = 0;
 
             $albaran->envio_nombre = $pedido->envio_nombre;
             $albaran->envio_apellidos = $pedido->envio_apellidos;
@@ -193,8 +198,7 @@ class ventas_megafacturador extends fs_controller {
             $albaran->envio_direccion = $pedido->envio_direccion;
             $albaran->envio_apartado = $pedido->envio_apartado;
 
-            if( is_null($albaran->codagente) )
-            {
+            if (is_null($albaran->codagente)) {
                 $albaran->codagente = $this->user->codagente;
             }
 
@@ -213,45 +217,80 @@ class ventas_megafacturador extends fs_controller {
             } else if (!$eje0->abierto()) {
                 $this->new_error_msg("El ejercicio está cerrado.");
             } else if ($albaran->save()) {
+                $trazabilidad = FALSE;
                 $continuar = TRUE;
-                $contador++;
+                $generar = TRUE;
+
                 $art0 = new articulo();
                 foreach ($pedido->get_lineas() as $l) {
-                    $n = new linea_albaran_cliente();
-                    $n->idlineapedido = $l->idlinea;
-                    $n->idpedido = $l->idpedido;
-                    $n->idalbaran = $albaran->idalbaran;
-                    $n->cantidad = $l->cantidad;
-                    $n->codimpuesto = $l->codimpuesto;
-                    $n->descripcion = $l->descripcion;
-                    $n->dtopor = $l->dtopor;
-                    $n->irpf = $l->irpf;
-                    $n->iva = $l->iva;
-                    $n->pvpsindto = $l->pvpsindto;
-                    $n->pvptotal = $l->pvptotal;
-                    $n->pvpunitario = $l->pvpunitario;
-                    $n->recargo = $l->recargo;
-                    $n->referencia = $l->referencia;
-
-                    if ($n->save()) {
-                        /// descontamos del stock
-                        if (!is_null($n->referencia)) {
-                            $articulo = $art0->get($n->referencia);
-                            if ($articulo) {
-                                $articulo->sum_stock($albaran->codalmacen, 0 - $l->cantidad);
-                            }
+                    //Si el articulo existe
+                    if(!is_null($l->referencia)){
+                    $articulo = $art0->get($l->referencia);
+                    //Si el producto no tiene control de stock que continue
+                        if ($articulo->controlstock) {
+                            $generar = TRUE;
+                            //Si el producto tiene control de stock y tiene stock suficiente que continue tambien
+                        } elseif (!$articulo->controlstock AND $articulo->stockfis >= $l->cantidad) {
+                            $generar = TRUE;
+                            //Pero si no entonces que no agregue esa linea
+                        } else {
+                            $generar = FALSE;
+                            $continuar = FALSE;
+                            $this->new_error_msg("¡No hay stock del artículo " . $l->referencia . " en el pedido <a href='".$pedido->url()."' target='_blank'>". $pedido->codigo . "</a> no se creará una linea para este artículo! ");
                         }
-                    } else {
-                        $continuar = FALSE;
-                        $this->new_error_msg("¡Imposible guardar la línea el artículo " . $n->referencia . "! ");
-                        break;
+                    }
+                    if($generar AND $continuar){
+                        $n = new linea_albaran_cliente();
+                        $n->idlineapedido = $l->idlinea;
+                        $n->idpedido = $l->idpedido;
+                        $n->idalbaran = $albaran->idalbaran;
+                        $n->cantidad = $l->cantidad;
+                        $n->codimpuesto = $l->codimpuesto;
+                        $n->descripcion = $l->descripcion;
+                        $n->dtopor = $l->dtopor;
+                        $n->irpf = $l->irpf;
+                        $n->iva = $l->iva;
+                        $n->pvpsindto = $l->pvpsindto;
+                        $n->pvptotal = $l->pvptotal;
+                        $n->pvpunitario = $l->pvpunitario;
+                        $n->recargo = $l->recargo;
+                        $n->referencia = $l->referencia;
+
+                        if ($n->save()) {
+                            /// descontamos del stock
+                            if (!is_null($n->referencia)) {
+                                $articulo = $art0->get($n->referencia);
+                                if ($articulo) {
+                                        $articulo->sum_stock($albaran->codalmacen, 0 - $l->cantidad, FALSE, $l->codcombinacion);
+                                }
+                            }
+                        } else {
+                            $continuar = FALSE;
+                            $this->new_error_msg("¡Imposible guardar la línea el artículo " . $n->referencia . "! ");
+                            break;
+                        }
+                        $albaran->neto += $l->pvptotal;
+                        $albaran->totaliva += ($l->pvptotal * $l->iva / 100);
+                        $albaran->totalirpf += ($l->pvptotal * $l->irpf / 100);
+                        $albaran->totalrecargo += ($l->pvptotal * $l->recargo / 100);
+                        if($l->irpf > $albaran->irpf)
+                        {
+                           $albaran->irpf = $l->irpf;
+                        }
                     }
                 }
+                //Validamos la información nueva del albarán
+                $albaran->neto = round($albaran->neto, FS_NF0);
+                $albaran->totaliva = round($albaran->totaliva, FS_NF0);
+                $albaran->totalirpf = round($albaran->totalirpf, FS_NF0);
+                $albaran->totalrecargo = round($albaran->totalrecargo, FS_NF0);
+                $albaran->total = $albaran->neto + $albaran->totaliva - $albaran->totalirpf + $albaran->totalrecargo;
 
-                if ($continuar) {
+                if ($continuar AND $generar) {
+                    if($albaran->save()){
                     $pedido->idalbaran = $albaran->idalbaran;
                     $pedido->fechasalida = $albaran->fecha;
-
+                        $contador++;
                     if (!$pedido->save()) {
                         $this->new_error_msg("¡Imposible vincular el " . FS_PEDIDO . " con el nuevo " . FS_ALBARAN . "!");
                         if ($albaran->delete()) {
@@ -260,9 +299,21 @@ class ventas_megafacturador extends fs_controller {
                             $this->new_error_msg("¡Imposible borrar el " . FS_ALBARAN . "!");
                         }
                     }
-                } else {
+                    }else{
+                        $this->new_error_msg('Ocurrio un error al intentar grabar el '.FS_ALBARAN.', hubo un problema con los artículos del '.FS_PEDIDO.' <a href="'.$pedido->url().'" target="_blank">'.$pedido->codigo.'</a> verifique el mismo e intente generar un albaran');
+                        $continuar=FALSE;
+                    }
+                } elseif(!$continuar AND !$generar) {
+                    $errores++;
                     if ($albaran->delete()) {
                         $this->new_error_msg("El " . FS_ALBARAN . " se ha borrado.");
+                    } else {
+                        $this->new_error_msg("¡Imposible borrar el " . FS_ALBARAN . "!");
+                    }
+                } elseif($continuar AND !$generar){
+                    $errores++;
+                    if ($albaran->delete()) {
+                        $this->new_error_msg("¡No hay stock para uno o varios artículos en el pedido <a href='".$pedido->url()."' target='_blank'>". $pedido->codigo . "</a> no se creará el ".FS_ALBARAN."! ");
                     } else {
                         $this->new_error_msg("¡Imposible borrar el " . FS_ALBARAN . "!");
                     }
@@ -271,7 +322,7 @@ class ventas_megafacturador extends fs_controller {
                 $this->new_error_msg("¡Imposible guardar el " . FS_ALBARAN . "!");
             }
         }
-        $this->new_message('Se procesaron correctamente '.$contador.' de '.$total.' pedidos.');
+        $this->new_message('Se procesaron correctamente ' . $contador . ' de ' . $total . ' pedidos y '.$errores.' no se procesaron por errores en stock o la información.');
     }
 
     /**
@@ -279,10 +330,13 @@ class ventas_megafacturador extends fs_controller {
      * Esta función genera las facturas del listado de albaranes pendientes
      * generados en base a las series de los albaranes y las fechas elegidas
      */
-    private function generar_facturas(){
-        $total = $this->total_pendientes_venta()['total'];
+    private function generar_facturas() {
+
+        $total_albaranes = $this->total_pendientes_venta();
+        $total = $total_albaranes['total'];
         $contador = 0;
         foreach ($this->pendientes_venta() as $albaran) {
+            $cliente = $this->cliente->get($albaran->codcliente);
             $contador++;
             $factura = new factura_cliente();
             $factura->apartado = $albaran->apartado;
@@ -321,15 +375,13 @@ class ventas_megafacturador extends fs_controller {
             $factura->totalrecargo = $albaran->totalrecargo;
             $factura->porcomision = $albaran->porcomision;
 
-            if( is_null($factura->codagente) )
-            {
+            if (is_null($factura->codagente)) {
                $factura->codagente = $this->user->codagente;
             }
 
             /// asignamos el ejercicio que corresponde a la fecha elegida
             $eje0 = $this->ejercicio->get_by_fecha($this->fecha_facturas_gen);
-            if($eje0)
-            {
+            if ($eje0) {
                $factura->codejercicio = $eje0->codejercicio;
                $factura->set_fecha_hora($this->fecha_facturas_gen, \date('H:i:s'));
             }
@@ -337,35 +389,24 @@ class ventas_megafacturador extends fs_controller {
             /// comprobamos la forma de pago para saber si hay que marcar la factura como pagada
             $forma0 = new forma_pago();
             $formapago = $forma0->get($factura->codpago);
-            if($formapago)
-            {
-               if($formapago->genrecibos == 'Pagados')
-               {
+            if ($formapago) {
+                if ($formapago->genrecibos == 'Pagados') {
                   $factura->pagada = TRUE;
-               }
-
-               $factura->vencimiento = Date('d-m-Y', strtotime($factura->fecha.' '.$formapago->vencimiento));
+                }
+                $factura->vencimiento = $formapago->calcular_vencimiento($factura->fecha, $cliente->diaspago);
             }
 
             $regularizacion = new regularizacion_iva();
 
-            if( !$eje0 )
-            {
-               $this->new_error_msg("Ejercicio no encontrado o está cerrado.");
-            }
-            else if( !$eje0->abierto() )
-            {
-               $this->new_error_msg("El ejercicio está cerrado.");
-            }
-            else if( $regularizacion->get_fecha_inside($factura->fecha) )
-            {
-               $this->new_error_msg("El ".FS_IVA." de ese periodo ya ha sido regularizado. No se pueden añadir más facturas en esa fecha.");
-            }
-            else if( $factura->save() )
-            {
-               $continuar = TRUE;
-               foreach($albaran->get_lineas() as $l)
-               {
+            if (!$eje0) {
+                $this->new_error_msg("Ejercicio no encontrado o está cerrado.");
+            } else if (!$eje0->abierto()) {
+                $this->new_error_msg("El ejercicio está cerrado.");
+            } else if ($regularizacion->get_fecha_inside($factura->fecha)) {
+                $this->new_error_msg("El " . FS_IVA . " de ese periodo ya ha sido regularizado. No se pueden añadir más facturas en esa fecha.");
+            } else if ($factura->save()) {
+                $continuar = TRUE;
+               foreach ($albaran->get_lineas() as $l) {
                   $n = new linea_factura_cliente();
                   $n->idalbaran = $l->idalbaran;
                   $n->idfactura = $factura->idfactura;
@@ -383,34 +424,30 @@ class ventas_megafacturador extends fs_controller {
                   $n->orden = $l->orden;
                   $n->mostrar_cantidad = $l->mostrar_cantidad;
                   $n->mostrar_precio = $l->mostrar_precio;
+                  $n->codcombinacion = $l->codcombinacion;
 
-                  if( !$n->save() )
-                  {
+                  if (!$n->save()) {
                      $continuar = FALSE;
-                     $this->new_error_msg("¡Imposible guardar la línea el artículo ".$n->referencia."! ");
+                     $this->new_error_msg("¡Imposible guardar la línea el artículo " . $n->referencia . "! ");
                      break;
                   }
                }
 
-               if($continuar)
-               {
+               if ($continuar) {
                   $albaran->idfactura = $factura->idfactura;
                   $albaran->ptefactura = FALSE;
-                  if( $albaran->save() )
-                  {
+                  if ($albaran->save()) {
                      $this->generar_asiento_cliente($factura);
                   } else {
-                     $this->new_error_msg("¡Imposible vincular el ".FS_ALBARAN." con la nueva factura!");
-                     if( $factura->delete() )
-                     {
+                     $this->new_error_msg("¡Imposible vincular el " . FS_ALBARAN . " con la nueva factura!");
+                     if ($factura->delete()) {
                         $this->new_error_msg("La factura se ha borrado.");
                      } else {
                         $this->new_error_msg("¡Imposible borrar la factura!");
                      }
                   }
                } else {
-                  if( $factura->delete() )
-                  {
+                  if ($factura->delete()) {
                      $this->new_error_msg("La factura se ha borrado.");
                   } else {
                      $this->new_error_msg("¡Imposible borrar la factura!");
@@ -419,9 +456,8 @@ class ventas_megafacturador extends fs_controller {
             } else {
                $this->new_error_msg("¡Imposible guardar la factura!");
             }
-
         }
-        $this->new_message('Se procesaron correctamente '.$contador.' de '.$total.' '.FS_ALBARANES);
+        $this->new_message('Se procesaron correctamente ' . $contador . ' de ' . $total . ' ' . FS_ALBARANES);
     }
 
     public function pendientes_venta() {
